Implement adding set rows to WorkoutActivity
WorkoutAdapter.kt
Move code for adding a set row to addSetRow(). This function goes through the adapterContents and checks if the current
WorkoutRowItem matches the exercise that we want to add a set to and that the next WorkoutRowItem does not match the
exercise. When that situation is found, we know that we have the last set for that exercise in the Recylerview.
 It then inserts a new WorkoutRowItem for a Set into the next adapterContents index and breaks.