Fix bug that caused EditText values to change when sets where added.
DefaultWorkoutViewHolder.kt
Add two MyEditTextListener parameters, one for lbs EditText and one for reps EditText.
If the view passed in has a tag of "Set Row", assign the MyEditTextListeners to the EditTexts.

WorkoutAdapter.kt
Assign tag of "Set Row" or "Exercise Row" to view. Pass two MyEditTextListeners to the DefaultWorkoutViewHolder for assigning to the
EditTexts in Set row items.
Add inner class MyEditTextListener. This class allows for using onTextChanged() to update the adapterContents list when changes are made
to the lbs or reps EditTexts. updateFunction() is called by the MyEditTextListeners in onBindViewHolder so that the listeners are tracking
the current EditText and not changing other rows.